+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## [0.3.0] - 2019-01-27
 - Added - Syntax highlighting should now be complete for .sl files.
-- Changed - Compiling should now work, although showing errors is not implemented yet.
+
+- Added - Compiling should now work, although showing errors is not implemented yet.
+
+- Added - Option to keep old rendered images: `rsl.images.keepHistory` (default: false).
+
+- Added - Option to name rendered images with a timestamp: `rsl.images.timestamp` (default: false).
+
+- Added - Option to pick format of rendered images: `rsl.images.format` (default: `PNG`).
 
 ## [0.2.0] - 2019-01-25
 - Added - Very basic support for compiling shaders.